Cập nhật

+++ backend/src/main/java/com/foodgo/backend/module/outlet/repository/OperatingHoursRepository.java
@@ -3,4 +3,5 @@ package com.foodgo.backend.module.outlet.repository;
 import com.foodgo.backend.module.outlet.entity.OperatingHours;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OperatingHoursRepository extends JpaRepository<OperatingHours, Long> {}
+public interface OperatingHoursRepository extends JpaRepository<OperatingHours, Integer> {
+}
