+++ backend/src/main/java/com/foodgo/backend/module/outlet/repository/OutletImageRepository.java
@@ -3,4 +3,5 @@ package com.foodgo.backend.module.outlet.repository;
 import com.foodgo.backend.module.outlet.entity.OutletImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OutletImageRepository extends JpaRepository<OutletImage, Long> {}
+public interface OutletImageRepository extends JpaRepository<OutletImage, Integer> {
+}
